Move additionalData() to public API
