+++ platforms/extensibility/plugin-development/src/test/groovy/org/gradle/plugin/devel/tasks/internal/ValidationProblemSerializationTest.groovy
@@ -18,6 +18,7 @@ package org.gradle.plugin.devel.tasks.internal
 
 import com.google.common.collect.HashMultimap
 import com.google.gson.Gson
+import org.gradle.api.problems.GeneralDataSpec
 import org.gradle.api.problems.Severity
 import org.gradle.api.problems.internal.AdditionalDataBuilderFactory
 import org.gradle.api.problems.internal.DefaultProblemReporter
@@ -25,7 +26,6 @@ import org.gradle.api.problems.internal.DeprecationData
 import org.gradle.api.problems.internal.DeprecationDataSpec
 import org.gradle.api.problems.internal.DocLink
 import org.gradle.api.problems.internal.GeneralData
-import org.gradle.api.problems.internal.GeneralDataSpec
 import org.gradle.api.problems.internal.GradleCoreProblemGroup
 import org.gradle.api.problems.internal.InternalProblemReporter
 import org.gradle.api.problems.internal.ProblemEmitter
