+++ platforms/ide/problems-api/src/main/java/org/gradle/api/problems/internal/AdditionalDataBuilderFactory.java
@@ -18,6 +18,8 @@ package org.gradle.api.problems.internal;
 
 import com.google.common.base.Function;
 import com.google.common.base.Preconditions;
+import org.gradle.api.problems.AdditionalDataSpec;
+import org.gradle.api.problems.GeneralDataSpec;
 import org.gradle.internal.service.scopes.Scope;
 import org.gradle.internal.service.scopes.ServiceScope;
