+++ platforms/ide/problems-api/src/main/java/org/gradle/api/problems/internal/DefaultGeneralData.java
@@ -17,6 +17,7 @@
 package org.gradle.api.problems.internal;
 
 import com.google.common.collect.ImmutableMap;
+import org.gradle.api.problems.GeneralDataSpec;
 
 import javax.annotation.Nullable;
 import java.io.Serializable;
