+++ platforms/ide/problems-api/src/main/java/org/gradle/api/problems/internal/DeprecationDataSpec.java
@@ -16,6 +16,8 @@
 
 package org.gradle.api.problems.internal;
 
+import org.gradle.api.problems.AdditionalDataSpec;
+
 /**
  * Specifies configuration options when creating a new DeprecationData instance.
  */
