+++ platforms/ide/problems-api/src/main/java/org/gradle/api/problems/internal/InternalProblemBuilder.java
@@ -17,6 +17,7 @@
 package org.gradle.api.problems.internal;
 
 import org.gradle.api.Action;
+import org.gradle.api.problems.AdditionalDataSpec;
 import org.gradle.api.problems.ProblemGroup;
 import org.gradle.api.problems.Severity;
