+++ platforms/ide/problems-api/src/main/java/org/gradle/api/problems/internal/InternalProblemSpec.java
@@ -17,6 +17,7 @@
 package org.gradle.api.problems.internal;
 
 import org.gradle.api.Action;
+import org.gradle.api.problems.AdditionalDataSpec;
 import org.gradle.api.problems.ProblemGroup;
 import org.gradle.api.problems.ProblemSpec;
 import org.gradle.api.problems.Severity;
@@ -25,18 +26,7 @@ import javax.annotation.Nullable;
 
 public interface InternalProblemSpec extends ProblemSpec {
 
-    /**
-     * Attaches additional data describing the problem.
-     * <p>
-     * Only the types listed for {@link AdditionalData} can be used as arguments, otherwise an invalid problem report will be created.
-     * <p>
-     * If not additional data was configured for this problem, then a new instance will be created. If additional data was already configured, then the existing instance will be used and the configuration will be applied to it.
-     *
-     * @param specType the type of the additional data configurer (see the AdditionalDataSpec interface for the list of supported types)
-     * @param config  The action configuring the additional data
-     * @return this
-     * @param <U> The type of the configurator object that will be applied to the additional data
-     */
+    @Override
     <U extends AdditionalDataSpec> InternalProblemSpec additionalData(Class<? extends U> specType, Action<? super U> config);
 
     /**
