+++ platforms/ide/problems-api/src/test/groovy/org/gradle/api/problems/internal/AdditionalDataBuilderFactoryTest.groovy
@@ -16,6 +16,7 @@
 
 package org.gradle.api.problems.internal
 
+import org.gradle.api.problems.AdditionalDataSpec
 import spock.lang.Specification
 
 import javax.annotation.Nullable
